% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjective_cooccurrence}
\alias{adjective_cooccurrence}
\title{Adjective co-occurrence}
\format{
A data frame with five variables, and two attributes:
attribute corpus_size contains the number of tokens in the whole corpus
attribute unique_jj contains the number of tokens tagged as adjectives
(jj.*) in the whole corpus

\describe{
\item{\code{word}}{first adjective}
\item{\code{collocate}}{second adjective}
\item{\code{o11}}{frequency of co-occurrence of \code{word} and \code{collocate}}
\item{\code{f1}}{frequency of \code{word}}
\item{\code{f2}}{frequency of \code{collocate}}
}
}
\usage{
adjective_cooccurrence
}
\description{
This data set includes counts for sentential co-occurrence of adjectives in
the Brown corpus, based on Justeson & Katz (1990). The data is extracted
from the same data set available in \code{?brown}. The tags used to extract
adjectives match "^jj.*".
}
\examples{

data(adjective_cooccurrence)
str(adjective_cooccurrence)

attributes(adjective_cooccurrence)$unique_jj
}
\references{
Justeson, John S., and Slava M. Katz. "Co-occurrences of antonymous adjectives and their contexts." Computational linguistics 17.1 (1991): 1-20.
}
\seealso{
\link{brown}
}
\keyword{datasets}
