% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_measures}
\alias{available_measures}
\title{Builtin measures}
\usage{
available_measures(stat = "")
}
\arguments{
\item{stat}{character, currently either "assoc" or "disp"}
}
\description{
Display names of built-in measures
}
\details{
The following collocation and association measures are currently available:

\describe{
\item{\code{ll}}{...}
\item{\code{mi}}{...}
\item{\code{mi2}}{...}
\item{\code{mi3}}{...}
\item{\code{fye}}{...}
\item{\code{poisson_pv}}{...}
\item{\code{rel_risk}}{...}
\item{\code{liddell}}{...}
\item{\code{gmean}}{...}
\item{\code{dice}}{...}
\item{\code{jaccard}}{...}
\item{\code{zscore}}{...}
\item{\code{zscore_cor}}{...}
\item{\code{chisq}}{...}
\item{\code{chisq_i}}{...}
\item{\code{chisq_h}}{...}
\item{\code{chisq_corr}}{...}
\item{\code{pearsonsid}}{...}
\item{\code{t_score}}{...}
\item{\code{cramer}}{...}
\item{\code{min_sens}}{...}
\item{\code{poisson_stirling}}{...}
\item{\code{odds_ratio}}{...}
\item{\code{delta_p1}}{...}
\item{\code{delta_p2}}{...}
}

The following dispersion measures are currently available:

\describe{
\item{\code{range}}{...}
\item{\code{maxmin}}{...}
\item{\code{engvall}}{...}
\item{\code{idf}}{...}
\item{\code{D}}{...}
\item{\code{U}}{...}
\item{\code{sd_pop}}{...}
\item{\code{cv_pop}}{...}
\item{\code{D_eq}}{...}
\item{\code{U_eq}}{...}
\item{\code{D2}}{...}
\item{\code{Um}}{...}
\item{\code{f_R}}{...}
\item{\code{S}}{...}
\item{\code{dc}}{...}
\item{\code{f_R_eq}}{...}
\item{\code{S_eq}}{...}
\item{\code{kld}}{...}
\item{\code{kld_norm}}{...}
\item{\code{dp}}{...}
\item{\code{dp_norm}}{...}
\item{\code{chisq}}{...}
\item{\code{D3}}{...}
\item{\code{Ur}}{...}
}

The following distance-based dispersion measures are currently available:

\describe{
\item{\code{arf}}{...}
\item{\code{awt}}{...}
\item{\code{f_awt}}{...}
\item{\code{ald}}{...}
\item{\code{f_ald}}{...}
\item{\code{washtell}}{...}
\item{\code{dwg}}{...}
\item{\code{dwg_norm}}{...}
}
}
