% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brown}
\alias{brown}
\title{Brown corpus}
\format{
A data frame with four factor variables: \code{text_id}, \code{genre},
\code{token}, \code{pos}; and two string attributes: \code{contents} and \code{readme}
}
\usage{
brown
}
\description{
The Brown corpus in tabular format tokenized and pos-tagged as distributed on
\url{https://www.nltk.org/nltk_data/}. Headings and sentence boundaries are
currently not preserved.
}
\details{
For documentation, see \url{http://korpus.uib.no/icame/brown/bcm.html}.
The the raw README and CONTENTS files are also included as attributes.
}
\examples{

head(brown)

class(attr(brown, "README"))
}
\keyword{datasets}
