% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{collexemes}
\alias{collexemes}
\alias{collexemes.data.frame}
\alias{collexemes.matrix}
\alias{collexemes.data.table}
\alias{collexemes.default}
\title{Association measures for collocation and collostruction}
\usage{
collexemes(.x, ...)

\method{collexemes}{data.frame}(.x, o11, f1, f2 = NULL, n = NULL, fun = "ll", flip = NULL, ...)

\method{collexemes}{matrix}(.x, o11, f1, f2 = NULL, n = NULL, fun = "ll", flip = NULL, ...)

\method{collexemes}{data.table}(.x, ...)

\method{collexemes}{default}(.x, o11, f1, f2 = NULL, n = NULL, fun = "ll", flip = NULL, ...)
}
\arguments{
\item{.x}{data.frame or list containing data}

\item{...}{further arguments to be passed to or from other methods}

\item{o11}{numeric vector with joint frequencies}

\item{f1}{numeric vector}

\item{f2}{numeric, if not provided, sum of o11 is used}

\item{n}{numeric, if not provided, sum of f1 is used}

\item{fun}{character vector for built-in measures (see Details).
alternatively a custom function, expression, or call. if names are supplied,
they are used in the output}

\item{flip}{character names of measures for which to flip the sign, intended
to signify negative association with two-sided measures}
}
\value{
matrix
}
\description{
Calculates association measures from vectors of frequencies and
joint frequencies.
}
\details{
TODO:
}
