% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{create_corpus}
\alias{create_corpus}
\title{Prepare corpus data for dispersion calculations}
\usage{
create_corpus(
  tokens,
  parts,
  freq = NULL,
  vocab = NULL,
  doc_ids = NULL,
  type = c("per_part", "raw"),
  cutoff = 0L,
  with_distance = TRUE,
  no_match = c("fail", "remove", "keep")
)
}
\arguments{
\item{tokens}{character vector with tokens}

\item{parts}{character vector}

\item{freq}{integer optional vector with counts}

\item{vocab}{character or factor optional vector with unique tokens}

\item{doc_ids}{character or factor optional vector with part ids}

\item{type}{input type, either "per_part" or "raw"}

\item{cutoff}{integer minimum frequency for each type}

\item{with_distance}{logical whether or not to calculate distances required
for distance measures}

\item{no_match}{character, "fail" (default): throws an error if tokens
contain NAs after creating an index. Typically, this happens when \code{vocab} is
given and doesn't contain all types in the corpus; "remove": NAs are removed,
"keep": treat NAs as separate type of token
error}
}
\value{
list of type "corpus"
}
\description{
From raw corpus or part-frequency list, create a list object containing
descriptive stats and indexed tokens.
}
\details{
\describe{
\item{\code{iparts}}{integer index of parts}
\item{\code{l}}{number of tokens in the input}
\item{\code{f}}{frequency per unique tokens}
\item{\code{i}}{integer index of tokens per parts}
\item{\code{j}}{integer index of parts per tokens}
\item{\code{v}}{frequency of tokens per part}
\item{\code{vocab}}{unique tokens}
\item{\code{sort_ids}}{sorting permutation of tokens for use in distance
based measures}
\item{\code{sizes}}{sizes of parts}
}
}
