% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal.R
\name{dp}
\alias{dp}
\title{Deviation of proportions}
\usage{
dp(tokens, parts, v, norm = TRUE)
}
\arguments{
\item{tokens}{character or factor of the same length as \code{v}}

\item{parts}{character or factor of the same length as \code{v}}

\item{v}{integer vector with per document frequencies}

\item{norm}{logical whether or not to normalize, see Details}
}
\value{
numeric
}
\description{
TODO:
}
\details{
TODO:
}
\examples{
n <- 50
v <- sample(1:100, n, replace = TRUE)
tokens <- sample(letters, n, replace = TRUE)
parts <- sample(LETTERS[1:3], n, replace = TRUE)
dp(tokens, parts, v)

}
