% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal.R
\name{dwg}
\alias{dwg}
\title{Word Growth Dispersion}
\usage{
dwg(tokens, corr = TRUE)
}
\arguments{
\item{tokens}{character raw corpus}

\item{corr}{logical whether or not to apply a geometric correction}
}
\value{
numeric
}
\description{
A distance-based dispersion measure
}
\details{
TODO:
}
\examples{
n <- 50
tokens <- sample(letters, n, replace = TRUE)
dwg(tokens)

}
