% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_compact.R
\name{ll_mini}
\alias{ll_mini}
\alias{ll_dt}
\title{Calculate G² (log-likelihood test) from vectors}
\usage{
ll_mini(f1, o11, n = sum(f1), f2 = sum(o11))

ll_dt(x, n = sum(x$f1), f2 = sum(x$o11), one_sided = TRUE, sorted = TRUE)
}
\arguments{
\item{f1}{numeric or integer with frequencies of token}

\item{o11}{numeric or integer with joint frequency}

\item{n}{numeric or integer with overall frequencies if not provided,
sum of f1 is taken}

\item{f2}{numeric or integer with frequencies of co-occurring token or construction
if not provided, sum of joint frequencies (o11) is taken}

\item{x}{data.table with named columns o11 and f1, optionally f2 and n}

\item{one_sided}{logical. should repulsion be indicated by a negative sign?}

\item{sorted}{logical. should output be sorted by association strength?}
}
\value{
numeric
}
\description{
Two functions that calculate loglikelihood association measures
for vectors of counts. Mostly for practice/demonstration purposes
}
\examples{
lol_test <- ll_mini(c(this = 41, is = 32, corpus = 41, data = 12), c(12, 1, 0, 3))
}
