% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prepare_coll_analysis}
\alias{prepare_coll_analysis}
\title{Prepare input for collocation or collostruction analysis}
\usage{
prepare_coll_analysis(
  joint,
  frequencies,
  type = c("collocation", "simple", "distinctive", "co-varying")
)
}
\arguments{
\item{joint}{data.frame joint frequencies of collocate 1 or word in
construction}

\item{frequencies}{data.frame corpus/sample frequencies of collocate 1 or
word}

\item{type}{character either of "collocation", "simple", "distinctive",
"co-varying", see details}
}
\value{
data.frame valid for coll_analysis
}
\description{
This is a convenience function to create valid input for \code{coll_analysis}
from raw frequency lists
}
\details{
TODO
}
\examples{
# TODO

}
