% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{available_measures}
\alias{available_measures}
\title{Builtin measures}
\usage{
available_measures(type)
}
\arguments{
\item{type}{character, either "assoc" or "disp"}
}
\description{
Display names of built-in association measures to be used in
}
