% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_measures}
\alias{available_measures}
\title{Builtin measures}
\usage{
available_measures(stat = NULL)
}
\arguments{
\item{stat}{character, currently either "assoc" or "disp"}
}
\description{
Display names of built-in measures
}
