% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{collexemes}
\alias{collexemes}
\title{Association measures for collocation and collostruction}
\usage{
collexemes(o11, f1, f2 = sum(o11), n = NULL, fun = "ll", flip = NULL)
}
\arguments{
\item{o11}{numeric vector with joint frequencies}

\item{f1}{numeric vector}

\item{f2}{numeric, if not provided, sum of o11 is used}

\item{n}{numeric, if not provided, sum of f1 is used}

\item{fun}{character vector for built-in measures (see Details).
alternatively a custom function, expression, or call. if names are supplied,
they are used in the output}

\item{flip}{character names of measures for which to flip the sign, intended
to signify negative association with two-sided measures}
}
\value{
matrix
}
\description{
Calculates association measures from vectors of frequencies and
joint frequencies.
}
\details{
TODO:
}
