% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\alias{dispersion.data.frame}
\alias{dispersion.table}
\alias{dispersion.default}
\title{Calculate word dispersion measures}
\usage{
dispersion(.x, ...)

\method{dispersion}{data.frame}(
  .x,
  tokens,
  parts,
  freq = NULL,
  fun = "dp_norm",
  type = c("per_part", "raw", "corpus"),
  cutoff = 0L,
  ...
)

\method{dispersion}{table}(.x, tokens, parts, fun = "dp_norm", cutoff = 0L, ...)

\method{dispersion}{default}(
  .x = NULL,
  tokens,
  parts,
  freq = NULL,
  fun = "dp.norm",
  type = c("per_part", "raw", "corpus"),
  cutoff = 0L,
  ...
)
}
\arguments{
\item{.x}{data.frame or list containing data}

\item{...}{further arguments to be passed to or from other methods}

\item{tokens}{character, factor, or expression specifying the data.frame
column or table dim}

\item{parts}{character, factor, or expression specifying the data.frame
column or table dim}

\item{freq}{integer counts}

\item{fun}{character or named list.}

\item{type}{input type, see details}

\item{cutoff}{integer minimum frequency for each type}
}
\description{
Calculate word dispersion measures
}
