% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\title{Calculate word dispersion measures}
\usage{
dispersion(tokens, parts, v, fun = "dp.norm", lexicon = NULL)
}
\arguments{
\item{tokens}{character or factor. token}

\item{parts}{character or factor.
part id where a part is usually text or text region.}

\item{v}{integer. counts}

\item{fun}{(named) character or list.}

\item{lexicon}{character. optional list of words to match. can be used
for clean up and speed up}
}
\description{
Calculate word dispersion measures
}
