% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\alias{dispersion.data.frame}
\alias{dispersion.table}
\alias{dispersion.default}
\title{Calculate word dispersion measures}
\usage{
dispersion(.x, ...)

\method{dispersion}{data.frame}(.x, tokens, parts, v = NULL, fun = "dp_norm", ...)

\method{dispersion}{table}(.x, tokens, parts, ...)

\method{dispersion}{default}(.x, tokens, parts, v, fun, ...)
}
\arguments{
\item{.x}{data.frame or list containing data}

\item{...}{further arguments to be passed to or from other methods}

\item{tokens}{character or factor; for table method, character specifying
dimname; for data.frame method, expression specifying the column}

\item{parts}{character or factor; for table method, character specifying
dimname; for data.frame method, expression specifying the column
part id where a part is usually text or text region.}

\item{v}{integer. counts}

\item{fun}{(named) character or list.}
}
\description{
Calculate word dispersion measures
}
