% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal.R
\name{dwg}
\alias{dwg}
\title{Word Growth Dispersion}
\usage{
dwg(cpos, f, size, corr = TRUE)
}
\arguments{
\item{cpos}{integer vector representing the indeces of a token}

\item{f}{integer count of a token}

\item{size}{integer corpus size}

\item{corr}{logical whether or not to apply a geometric correction, see
Details}
}
\value{
numeric
}
\description{
A distance-based dispersion measure
}
\details{
TODO:
}
\examples{
dwg(c(41, 32, 41, 12), f = 4, size = 100)

}
