% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal.R
\name{ll}
\alias{ll}
\title{Calculate G² (log-likelihood test) from vectors}
\usage{
ll(f1, o11, n = sum(f1), f2 = sum(o11), one_sided = TRUE)
}
\arguments{
\item{f1}{numeric or integer with frequencies of token}

\item{o11}{numeric or integer with joint frequency}

\item{n}{numeric or integer with overall frequencies if not provided,
sum of f1 is taken}

\item{f2}{numeric or integer with frequencies of co-occurring token or
construction if not provided, sum of joint frequencies (o11) is taken}

\item{one_sided}{logical if \code{TRUE}, repulsion is indicated by a negative sign}
}
\value{
numeric
}
\description{
Calculate loglikelihood association measures for vectors of counts.
}
\examples{
ll(c(this = 41, is = 32, corpus = 41, data = 12), c(12, 1, 0, 3))

}
