% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_compact.R
\name{ll_mini}
\alias{ll_mini}
\title{Calculate G² (log-likelihood test) from vectors}
\usage{
ll_mini(f1, o11, n = sum(f1), f2 = sum(o11))
}
\arguments{
\item{f1}{numeric or integer with frequencies of token}

\item{o11}{numeric or integer with joint frequency}

\item{n}{numeric or integer with overall frequencies if not provided,
sum of f1 is taken}

\item{f2}{numeric or integer with frequencies of co-occurring token or construction
if not provided, sum of joint frequencies (o11) is taken}
}
\value{
numeric
}
\description{
Two functions that calculate loglikelihood association measures
for vectors of counts. Mostly for practice/demonstration purposes
}
\examples{
lol_test <- ll_mini(c(thes = 41, is = 32, corpus = 41, data = 12), c(12, 1, 0, 3))

}
